refactor: extract BudgetState assembly into domain and hook layers
- Add createBudgetState pure function in core/budget.ts
- Add useBudgetState hook in app/hooks/
- Move useSectionData to app/hooks/ directory
- Simplify page.tsx to use useBudgetState
